Format all Python files using black formatter.
This is necessary to reduce the future changes made by black formatter that executes automatically.

This will be useful to reduce conflicts when merging changes.

After that, when commiting or merging code, the developers can focus only in yours changes instead to see too all changes made by black formatter.